Refactored project structure
* Added src folder

* Atoms and molecules mixed

* Pages renamed to screens

* Moved components, adapters and others to src
